EDU-4830 Merchant Article API refractoring

* EDU-4830 Merchant Article API refractoring

Approved-by: Shallu Mangla

+++ digital-education-explore/src/main/java/com/paytm/digital/education/explore/response/dto/articles/ArticleResponseDTO.java
@@ -19,7 +19,7 @@ public class ArticleResponseDTO {
     private String articleTitle;
 
     @JsonProperty("article_id")
-    private Integer articleId;
+    private Long articleId;
 
     @JsonProperty("article_url")
     private String articleUrl;
@@ -31,7 +31,7 @@ public class ArticleResponseDTO {
     private String articleDescription;
 
     @JsonProperty("exam_id")
-    private Integer examId;
+    private Long examId;
 
     @JsonProperty("stream")
     private String stream;
